[#114] :bug: fix: 테스트 코드 수정 및 통합
