--- src/test/java/team/seventhmile/tripforp/domain/plan/repository/PlanGetDetailRepositoryTest.java
@@ -1,4 +1,0 @@
-package team.seventhmile.tripforp.domain.plan.repository;
-
-public class PlanGetDetailRepositoryTest {
-}
